fixed inconsistent array implementation in quarry-data.js
Due to a misunderstanding of Ember.Array, we were creating Ember Enumerable objects via "Em.A()". In fact, Ember extends the Array prototype to provide this. From the docs:

“Array is extended to implement the Ember.Enumerable, Ember.MutableEnumerable, Ember.MutableArray and Ember.Array interfaces. This polyfills ECMAScript 5 array methods in browsers that do not implement them, adds convenience methods and properties to built-in arrays, and makes array mutations observable.”

http://emberjs.com/guides/configuring-ember/disabling-prototype-extensions/

Ember.A() is only useful if “prototype extensions” are disabled, in which case the array literals will not be Ember Enumerable objects.

Also corrected this mistaken usage in other areas of the codebase.