Expose asset id and vm vnc port in asset details
Prior to this commit the asset id value and vm vnc port information was
not exposed in the asset details template. Not having these values
quickly accessible has resulted in a number of user complaints.

We now expose the asset id in the general tab and the vnc port in the vm
tab of assets with corresponding vm records.

commit 7fd166a7e910c497bdef0193b735b706a5e533eb
Author: Richard Marshall <richard.marshall@ask.com>
Date:   Mon Sep 15 16:44:10 2014 -0700

    Add vnc information to asset vm tab