Display related asset in job views
This commit adds logic to lookup the FQDN of an asset that is acted upon
by a job. This is done either by extracting the FQDN from the job
args/kwargs or by making a call to the Asset history API with an
asset_id.

The fetched FQDN is then displayed on the job list and job details views
as a link to the corresponding asset details view.